requirements.txt
python-telegram-bot
influxdb
telethon
